add akka implementation and some refactoring
