+++ implementations/monix/src/main/scala/sttp/client3/impl/monix/MonixServerSentEvents.scala
@@ -1,0 +1,8 @@
+package sttp.client3.impl.monix
+
+import monix.reactive.Pipe
+import sttp.client3.sse.ServerSentEvent
+
+object MonixServerSentEvents {
+  def decodeSSE: Pipe[Array[Byte], ServerSentEvent] = ???
+}
