+++ implementations/monix/src/test/scala/sttp/client3/impl/monix/MonixStreamingTest.scala
@@ -19,5 +19,8 @@ abstract class MonixStreamingTest extends StreamingTest[Task, MonixStreams] {
     stream.toListL
       .map(bs => new String(bs.toArray.flatten, "utf8"))
 
-  override def sseConsumer(stream: Observable[Array[Byte]]): Task[List[ServerSentEvent]] = ???
+  override def sseConsumer(stream: Observable[Array[Byte]]): Task[List[ServerSentEvent]] =
+    stream.pipeThrough(MonixServerSentEvents.decodeSSE).foldLeftL(List.empty[ServerSentEvent]) { case (list, event) =>
+      list :+ event
+    }
 }
